SVT-Assignment-main/Question_2/Task_04/src/test/java/GradeCalculatorTest.java
import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertThrows;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.CsvSource;

public class GradeCalculatorTest {

    @ParameterizedTest
    @CsvSource({
            "95, A",
            "85, B",
            "75, C",
            "65, D",
            "55, F"
    })
    void testGetGrade(int score, String expectedGrade) {
        assertEquals(expectedGrade, GradeCalculator.getGrade(score));
    }

    @Test
    void testInvalidScores() {
        assertThrows(IllegalArgumentException.class, () -> GradeCalculator.getGrade(-1));
        assertThrows(IllegalArgumentException.class, () -> GradeCalculator.getGrade(101));
    }
}